// src/types/JsonEditor.types.ts
export interface DefaultFileElementProps {
	value: {
		type: string;
		format: string | null;
		filename: string | null;
		base64: string | null;
	} | null;
	readModeValue: {
		type: string;
		format: string | null;
		filename: string | null;
		base64: string | null;
	} | null;
	path: string;
	format?: string;
}